added support for external capture plugins - and using avcodec as a plugin.c. The ragdoll API is potentially usable now, but still really limited. Enabled SQL requests by default using sqlite. Note that you'll need the sqlite dll to use this. MySQL should still be usable, but I didn't try. MySQL requires -DUSE_MYSQL to compile it, and a dll and -mysql argument to enable it. Fixed nacl. NPFTE plugin now invokes an exe to run the game rather than running the game within the browser. externvalue builtin now accepts & prefix to return a pointer instead. Fixed vector autocvars. uri_get, bufstr_add, bufstr_free, now functional. QC debugger can now show asm if line numbers are not available. Added support for QC watchpoints. Use the watchpoint command. gl_specular now give specular even without rtlights, thankfully not as blatently, but its there. android will not crash due to supported audio formats, and gles2 can be selected via a cvar (requires full FTEDroidActivity/program restart).
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4152 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ plugins/berkelium/plugapi.cpp
@@ -8,6 +8,8 @@
 
 #include <string>
 
+qboolean inited;
+
 class decctx
 {
 public:
@@ -188,6 +190,13 @@ static void *Dec_Create(char *medianame)
 	else
 		return NULL;
 
+	if (!inited)
+	{
+		//linux lags behind and apparently returns void, so don't bother checking return values on windows, cos I'm lazy.
+		Berkelium::init(Berkelium::FileString::empty());
+		inited = qtrue;
+	}
+
 	decctx *ctx = new decctx();
 
 	Berkelium::Context* context = Berkelium::Context::create();
@@ -366,22 +375,23 @@ static void Dec_ChangeStream(void *vctx, char *newstream)
 
 static bool Dec_Init(void)
 {
-	//linux lags behind and apparently returns void.
-	Berkelium::init(Berkelium::FileString::empty());
 	return true;
 }
 
 static qintptr_t Dec_Tick(qintptr_t *args)
 {
 	//need to keep it ticking over, if any work is to be done.
-	Berkelium::update();
+	if (inited)
+		Berkelium::update();
 	return 0;
 }
 
 static qintptr_t Dec_Shutdown(qintptr_t *args)
 {
 	//force-kill all.
-	Berkelium::destroy();
+	if (inited)
+		Berkelium::destroy();
+	inited = qfalse;
 	return 0;
 }
